feat(web): use trpc for module modal
